pyproject.toml
[tool.poetry]
name = "input-insight"
version = "0.1.0"
description = ""
authors = ["Frederic <zetachow@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pika = "^1.3.2"
spacy = "3.7.6"
flask = "^3.0.3"
langdetect = "^1.0.9"
pyyaml = "^6.0.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
